+++ src/shapes/fabricObject.class.ts
@@ -6,6 +6,7 @@ import { FabricObjectSVGExportMixin } from '../mixins/object.svg_export';
 import { applyMixins } from '../util/applyMixins';
 import { fabric } from '../../HEADER';
 import { FabricObject } from './object.class';
+import { FabricObjectObjectAnimationMixin } from '../mixins/object_animation.mixin';
 
 // TODO somehow we have to make a tree-shakeable import
 
@@ -14,13 +15,14 @@ applyMixins(InteractiveFabricObject, [
   FabricObjectObjectStackingMixin,
   FabricObjectObjectStraighteningMixin,
   FabricObjectSVGExportMixin,
+  FabricObjectObjectAnimationMixin,
 ]);
 
 // export interface InteractiveFabricObject
 //   extends FabricObjectAncestryMixin,
 //     FabricObjectObjectStackingMixin,
 //     FabricObjectObjectStraighteningMixin,
-//     FabricObjectSVGExportMixin {}
+//     FabricObjectSVGExportMixin,FabricObjectObjectAnimationMixin {}
 
 export { InteractiveFabricObject as FabricObject };
